% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{PANTHER-package}
\alias{PANTHER-package}
\title{PANTHER}
\description{
PANTHER database annotations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r.acidgenomics.com/packages/panther/}
  \item \url{https://github.com/acidgenomics/r-panther/}
  \item Report bugs at \url{https://github.com/acidgenomics/r-panther/issues/}
}

}
\author{
\strong{Maintainer}: Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})

Authors:
\itemize{
  \item Michael Steinbaugh \email{mike@steinbaugh.com} (\href{https://orcid.org/0000-0002-2403-2221}{ORCID})
}

Other contributors:
\itemize{
  \item Acid Genomics [copyright holder, funder]
}

}
\keyword{internal}
