% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerators.R
\name{PANTHER}
\alias{PANTHER}
\title{PANTHER database annotations}
\usage{
PANTHER(organism, release = NULL,
  BPPARAM = BiocParallel::SerialParam(progressbar = TRUE))
}
\arguments{
\item{organism}{\code{character(1)}.
Full Latin organism name (e.g. \code{"Homo sapiens"}).}

\item{release}{\code{character(1)} or \code{NULL}.
PANTHER release version. If \code{NULL}, defaults to current release. Consult
the PANTHER website for a list of release versions available from the FTP
server (e.g. \code{"14.0"}).}

\item{BPPARAM}{\code{bpparamClass}.
BiocParallel parameter to specify the desired processor configuration.\cr
We recommend using one of the following:
\itemize{
\item \link[BiocParallel:bpparam]{bpparam}.
\item \link[BiocParallel:SerialParam]{SerialParam}.
\item \link[BiocParallel:MulticoreParam]{MulticoreParam}.
}}
}
\value{
\code{PANTHER}.
}
\description{
\href{http://www.pantherdb.org}{PANTHER} gene ontology definitions. PANTHER stands
for \strong{P}rotein \strong{AN}alysis \strong{TH}rough \strong{E}volutionary \strong{R}elationships.
}
\note{
Updated 2019-08-27.
}
\section{Suported organisms}{

\itemize{
\item \emph{Caenorhabditis elegans}
\item \emph{Drosophila melanogaster}
\item \emph{Homo sapiens}
\item \emph{Mus musculus}
}
}

\examples{
options(acid.test = TRUE)
x <- PANTHER(
    organism = "Homo sapiens",
    BPPARAM = BiocParallel::SerialParam()
)
summary(x)
}
